+++ src/utilities/dataHelpers.h
@@ -23,10 +23,10 @@ void readImdbTitleIdColumnFromBasicsTable(const std::string& filePath, uint32_t*
 void readImdbStartYearColumnFromBasicsTable(const std::string& filePath, int* data);
 void readImdbStartYearColumnFromBasicsTable(const std::string& filePath, uint32_t* data);
 
-void readImdbTitleIdColumnFromPrincipalsTable(const std::string& filePath, uint32_t* data);
+void readImdbTitleIdColumnFromPrincipalsTable(const std::string& filePath, uint64_t* data);
 void readImdbPersonIdColumnFromPrincipalsTable(const std::string& filePath, uint32_t* data);
 
-void readImdbTitleIdColumnFromAkasTable(const std::string& filePath, int* data);
+void readImdbTitleIdColumnFromAkasTable(const std::string& filePath, uint64_t* data);
 void readImdbTitleIdColumnFromAkasTable(const std::string& filePath, int64_t* data);
 
 template <typename T>
